+++ src/Microsoft.Sbom.Common/Conformance/Interfaces/IConformanceStandardErrorType.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Microsoft.Sbom.Parsers.Spdx30SbomParser.ConformanceStandard.Interfaces;
+namespace Microsoft.Sbom.Parsers.Spdx30SbomParser.Conformance.Interfaces;
 
 public interface IConformanceStandardErrorType
 {
