+++ test/Microsoft.Sbom.Api.Tests/Workflows/SbomParserBasedValidationWorkflowTests.cs
@@ -22,8 +22,8 @@ using Microsoft.Sbom.Api.Workflows;
 using Microsoft.Sbom.Api.Workflows.Helpers;
 using Microsoft.Sbom.Common;
 using Microsoft.Sbom.Common.Config;
-using Microsoft.Sbom.Common.ConformanceStandard;
-using Microsoft.Sbom.Common.ConformanceStandard.Enums;
+using Microsoft.Sbom.Common.Conformance;
+using Microsoft.Sbom.Common.Conformance.Enums;
 using Microsoft.Sbom.Contracts;
 using Microsoft.Sbom.Contracts.Enums;
 using Microsoft.Sbom.Extensions;
